Clean up wallet meta data and merge code in deterministic.go into wallet.go
Don't add wallet folder info in wallet meta info, it's added and has never been used
